Make map.[old]ypos ints instead of floats
So... I did see that map.ypos was a float when I added over-30-FPS mode,
because map.oldypos wasn't there before... I'm guessing that I kind of
just ignored it at the time. But, c'mon, map.ypos and map.oldypos are
always treated as ints, so there's literally no reason for them to be
actually floats in reality. I didn't even know they were anything other
than ints until I checked Map.h.

+++ desktop_version/src/Map.h
@@ -107,8 +107,8 @@ public:
 
     //Special tower stuff
     bool towermode;
-    float ypos;
-    float oldypos;
+    int ypos;
+    int oldypos;
     int cameramode;
     int cameraseek, cameraseekframe;
     int resumedelay;
